src/generator/requirements.txt
# Generator Module Requirements
# RAG mit FAISS und Alibaba Cloud Qwen

# LLM API
requests>=2.31.0

# Embeddings
sentence-transformers>=2.2.2
torch>=2.0.0

# Vector Store
faiss-cpu>=1.7.4
# oder für GPU: faiss-gpu>=1.7.4

# Database
pymongo>=4.5.0

# Utilities
numpy>=1.24.0

